+++ NEWS.md
@@ -1,6 +1,6 @@
 # Changes in mirt 1.28
 
-- `M2()`, `itemfit()`, and `fscores()` gain a `na.rm` logical to remove rows of missing data internally
+- `M2()`, `itemfit()`, `SIBTEST()`, and `fscores()` gain an `na.rm` logical to remove rows of missing data
 
 - `fscores()` gains a `append_response.pattern` logical to indicate whether response patterns via the
   `response.pattern` input should be appended to the factor score results